% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_lite.R
\name{clr_imputed}
\alias{clr_imputed}
\title{Resample random values, perform CLR over each iteration and return the median result.}
\usage{
clr_imputed(vec, method = "logunif", replicates = 1000)
}
\arguments{
\item{vec}{A vector of compositional data that may contain zeroes.}

\item{method}{The method for zero imputation. One of "logunif", "unif" or "const".}

\item{replicates}{A positive integer. Default is 1000. Controls how many replicates the median should be taken over.}
}
\value{
a vector of CLR-transformed data
}
\description{
See \code{clr_lite}.
}
\seealso{
\code{\link{clr_lite}}
}
