% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{coef_to_wide}
\alias{coef_to_wide}
\title{Take all coefficients and prepare them for wide format.}
\usage{
coef_to_wide(x)
}
\description{
Take all coefficients and prepare them for wide format.
}
