% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{compare_to_null}
\alias{compare_to_null}
\title{Fit a null model for a linear mixed effect model
s}
\usage{
compare_to_null(f.null, f.full)
}
\description{
Fit a null model for a linear mixed effect model
s
}
