% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidebook_data.R
\docType{data}
\name{counts}
\alias{counts}
\title{The curated and aggregated genus-level count table from the schizophrenia study for demonstration}
\format{
A data.frame object with 162 rows, genera, and  171 columns, samples.
}
\source{
\url{https://doi.org/10.1038/s41467-020-15457-9}
}
\usage{
counts
}
\description{
A genus-level count table for instructional purposes. Accessed using the curatedMetagenomicData R package
}
\keyword{datasets}
