% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asymptotic_alpha.R
\name{div_applier}
\alias{div_applier}
\title{Functional wrapper to get the output of the wonderful iNext library in the format I use to pipe into ggplot2.}
\usage{
div_applier(x)
}
\description{
Functional wrapper to get the output of the wonderful iNext library in the format I use to pipe into ggplot2.
}
