% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{fw_fit}
\alias{fw_fit}
\title{run a model on every row of your data. Made with microbiome data in mind.}
\usage{
fw_fit(
  x,
  f,
  metadata,
  verbose = T,
  get_CI = T,
  format = "wide",
  model = "lm",
  order = "fact",
  ...
)
}
\description{
run a model on every row of your data. Made with microbiome data in mind.
}
\examples{

metadata = data.frame(a = sample(letters[1:3], ncol(mtcars), replace=T),
                      b = sample(letters[4:6], ncol(mtcars), replace=T),
                      c = rnorm(ncol(mtcars)))

 fw_fit(x = mtcars, f = x ~ a + b, metadata = metadata, format = "fits", model = "lm", get_CI = T, order = "fact")

}
