% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{fw_glm}
\alias{fw_glm}
\title{lm wrapper around fw_fit.}
\usage{
fw_glm(
  x,
  f,
  metadata,
  verbose = T,
  get_CI = T,
  format = "wide",
  order = "fact",
  ...
)
}
\description{
lm wrapper around fw_fit.
}
