% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{fw_glmer}
\alias{fw_glmer}
\title{lmer wrapper around fw_fit.}
\usage{
fw_glmer(
  x,
  f,
  metadata,
  verbose = T,
  get_CI = T,
  format = "wide",
  order = "c",
  adjust.method = "BH",
  ...
)
}
\description{
lmer wrapper around fw_fit.
}
