% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidebook_data.R
\docType{data}
\name{GBMs}
\alias{GBMs}
\title{The curated Gut-Brain modules count table from the schizophrenia study for demonstration}
\format{
A data.frame object with 43 rows, Gut-Brain modules, and  171 columns, samples.
}
\source{
\url{https://doi.org/10.1038/s41467-020-15457-9}
}
\usage{
GBMs
}
\description{
A Gut-Brain modules count table for instructional purposes. Accessed using the curatedMetagenomicData R package
}
\keyword{datasets}
