% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_asymptotic_alpha.R
\name{get_asymptotic_alpha}
\alias{get_asymptotic_alpha}
\title{Functional wrapper to get the output of the wonderful iNext library in the format I use to pipe into ggplot2.}
\usage{
get_asymptotic_alpha(species, groups = c(), verbose = TRUE)
}
\description{
Functional wrapper to get the output of the wonderful iNext library in the format I use to pipe into ggplot2.
}
