% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_lite.R
\name{impute_zeroes}
\alias{impute_zeroes}
\title{Replace zeroes with non-zero values in order to perform a CLR-transformation}
\usage{
impute_zeroes(vec, method = "logunif")
}
\arguments{
\item{vec}{A vector of compositional data that may contain zeroes.}

\item{method}{The method for zero imputation. One of "logunif", "unif" or "const".}
}
\value{
A vector with all the zeroes replaced with non-zero values.
}
\description{
See \code{\link{clr_lite}}.
}
\seealso{
\code{\link{clr_lite}}
}
