% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{interpret_order}
\alias{interpret_order}
\title{Go from vector to individual booleans}
\usage{
interpret_order(x)
}
\arguments{
\item{x}{A string of input flags}
}
\value{
A list of orders
}
\description{
Go from vector to individual booleans
}
