% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{make_output}
\alias{make_output}
\title{Make a container list based on the order}
\usage{
make_output(x)
}
\arguments{
\item{x}{the order list output from `interpret_order`.}
}
\description{
Make a container list based on the order
}
