% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidebook_data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{A snippet of the metadata from the schizophrenia study}
\format{
A data.frame object with 171 rows, samples, and 10 columns, denoting sample ID, schizophrenia status, sex, BMI, smoking status and some additional information from the original metadata.
}
\source{
\url{https://doi.org/10.1038/s41467-020-15457-9}
}
\usage{
metadata
}
\description{
Contains the per-sample information for the samples from the schizophrenia study. Accessed using the curatedMetagenomicData R package
}
\keyword{datasets}
