% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.adjust.resample.R
\name{p.adjust.resamp}
\alias{p.adjust.resamp}
\title{Adjust p-values using resampling}
\usage{
p.adjust.resamp(p, method = "BH", n.samples = 100)
}
\arguments{
\item{p}{numeric vector of p-values}

\item{method}{correction method, a character string. Can be abbreviated. see `p.adjust.methods`}

\item{n.samples}{Integer. How many times should p-values be resampled}
}
\value{
A numeric vector of corrected p-values (of the same length as p, with names copied from p).
}
\description{
Wrapper around `p.adjust`, for simple use will do the same.
}
