% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.adjust.resample.R
\name{p_resamp_q}
\alias{p_resamp_q}
\title{Compute a single q-value through resampling.}
\usage{
p_resamp_q(x, p, n, method)
}
\arguments{
\item{x}{the p-value to be corrected}

\item{p}{numeric vector of the other p-values to sample from.}

\item{method}{correction method, a character string. Can be abbreviated. see `p.adjust.methods`}

\item{n.samples}{Integer. How many times should p-values be resampled}
}
\value{
A corrected p-value.
}
\description{
Compute a single q-value through resampling.
}
