% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_aitchison_PERMANOVA.R
\name{pairwise_aitchison_PERMANOVA}
\alias{pairwise_aitchison_PERMANOVA}
\title{Perform pairwise PERMANOVAs as a post-hoc}
\usage{
pairwise_aitchison_PERMANOVA(
  clr.samples,
  groups,
  relevant.comparisons,
  adjust.p = T,
  adj.method = "holm"
)
}
\description{
Perform pairwise PERMANOVAs as a post-hoc
}
