% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_DA_tester.R
\name{pairwise_DA_tester}
\alias{pairwise_DA_tester}
\title{Perform pairwise t-tests}
\usage{
pairwise_DA_tester(
  clr,
  groups,
  comparisons,
  verbose = TRUE,
  parametric = T,
  ignore.posthoc = F,
  posthoc.method = "BH",
  paired.test = FALSE
)
}
\description{
Perform pairwise t-tests
}
