% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_wise_modeller.R
\name{unroll_coefs}
\alias{unroll_coefs}
\title{Take all coefficients and prepare them for wide format. Called by `coef_to_wide`.}
\usage{
unroll_coefs(x)
}
\description{
Take all coefficients and prepare them for wide format. Called by `coef_to_wide`.
}
